third_party/blink/web_tests/fast/css/parsing-text-rendering-expected.txt

This tests checks that all of the input values for text-rendering parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test("text-rendering: auto;") is "auto"
PASS test("text-rendering: optimizespeed;") is "optimizespeed"
PASS test("text-rendering: optimizelegibility;") is "optimizelegibility"
PASS test("text-rendering: geometricprecision;") is "geometricprecision"
PASS test("text-rendering: OptIMizESpEEd;") is "optimizespeed"
PASS test("text-rendering: auto auto;") is ""
PASS test("text-rendering: optimizeCoconuts;") is ""
PASS test("text-rendering: 15;") is ""
PASS successfullyParsed is true

TEST COMPLETE
